Move UnknownOptions in consistency with the changes in master.
