+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/lookup/proto/ProtoToJavaTypeMapper.java
@@ -37,7 +37,7 @@ import static com.google.common.base.Strings.isNullOrEmpty;
 import static com.google.common.collect.Lists.newLinkedList;
 import static io.spine.gradle.compiler.util.JavaCode.getOuterClassName;
 import static io.spine.option.OptionsProto.TYPE_URL_PREFIX_FIELD_NUMBER;
-import static io.spine.protobuf.UnknownOptions.getUnknownOptionValue;
+import static io.spine.option.UnknownOptions.getUnknownOptionValue;
 
 /**
  * Maps Protobuf message types from a {@code .proto} file to the corresponding Java classes.
